Proof of concept with tooltips without a wrapper component?
